+++ ios/ios/src/Users/CMUser.m
@@ -43,7 +43,8 @@ static CMWebService *webService;
 
 @implementation CMUser
 
-@synthesize userId;
+@synthesize userId = _userId;
+@synthesize email = _email;
 @synthesize password;
 @synthesize token;
 @synthesize tokenExpiration;
@@ -51,7 +52,6 @@ static CMWebService *webService;
 @synthesize isDirty;
 @synthesize services;
 @synthesize username;
-@synthesize email;
 
 + (NSString *)className {
     return NSStringFromClass([self class]);
@@ -107,8 +107,8 @@ static CMWebService *webService;
         }
         token = [coder decodeObjectForKey:@"token"];
         tokenExpiration = [coder decodeObjectForKey:@"tokenExpiration"];
-        userId = [coder decodeObjectForKey:@"userId"];
-        email = [coder decodeObjectForKey:@"email"];
+        _userId = [coder decodeObjectForKey:@"userId"];
+        _email = [coder decodeObjectForKey:@"email"];
         username = [coder decodeObjectForKey:@"username"];
         services = [coder decodeObjectForKey:@"services"];
         if (!webService) {
@@ -156,7 +156,9 @@ static CMWebService *webService;
         id newValue = [change objectForKey:NSKeyValueChangeNewKey];
         if (![oldValue isEqual:newValue]) {
             // Only apply the change if a change was actually made.
-            NSLog(@"Detected change for property %@. Old value was \"%@\", new value is \"%@\"", keyPath, oldValue, newValue);
+            #ifdef DEBUG
+                NSLog(@"Detected change for property %@. Old value was \"%@\", new value is \"%@\"", keyPath, oldValue, newValue);
+            #endif
             isDirty = YES;
         }
     }
@@ -187,8 +189,8 @@ static CMWebService *webService;
         return NO;
     }
 
-    if (userId) {
-        return ([[object userId] isEqualToString:userId] && [[object password] isEqualToString:password]);
+    if (_userId) {
+        return ([[object userId] isEqualToString:_userId] && [[object password] isEqualToString:password]);
     } else if (username) {
         return ([[object username] isEqualToString:username] && [[object password] isEqualToString:password]);
     } else {
@@ -240,6 +242,19 @@ static CMWebService *webService;
     isDirty = NO;
 }
 
+
+- (NSString *)userId {
+    @synchronized(self) {
+        return _email;
+    }
+}
+
+- (void)setUserId:(NSString *)userId {
+    @synchronized(self) {
+        _email = userId;
+    }
+}
+
 #pragma mark - Remote user account and session operations
 
 - (BOOL)isCreatedRemotely {
@@ -249,8 +264,6 @@ static CMWebService *webService;
 
 - (void)save:(CMUserOperationCallback)callback {
     [webService saveUser:self callback:^(CMUserAccountResult result, NSDictionary *responseBody) {
-        NSLog(@"Response Body: %@", responseBody);
-        NSLog(@"Profile: %@", [responseBody objectForKey:@"profile"]);
         [self copyValuesFromDictionaryIntoState:responseBody];
         if (callback) {
             callback(result, [NSDictionary dictionary]);
@@ -274,9 +287,6 @@ static CMWebService *webService;
             objectId = [userProfile objectForKey:CMInternalObjectIdKey];
             
             self.services = [[responseBody objectForKey:@"profile"] objectForKey:@"__services__"];
-            
-            NSLog(@"Response Body: %@", responseBody);
-            NSLog(@"User Profile: %@", userProfile);
 
             if (!self.isDirty) {
                 // Only bring the changes from the server into the object state if there weren't local modifications.
@@ -335,28 +345,8 @@ static CMWebService *webService;
     }];
 }
 
-//TODO: Update so it works
 - (void)changePasswordTo:(NSString *)newPassword from:(NSString *)oldPassword callback:(CMUserOperationCallback)callback {
-    [webService changePasswordForUser:self
-                           oldPassword:oldPassword
-                           newPassword:newPassword
-                              callback:^(CMUserAccountResult result, NSDictionary *responseBody) {
-                                  if (result == CMUserAccountPasswordChangeSucceeded) {
-                                      self.password = newPassword;
-
-                                      // Since the password change succeeded, the user needs to be logged back
-                                      // in again to get a new session token since the old one has been expired.
-                                      [self loginWithCallback:^(CMUserAccountResult resultCode, NSArray *messages) {
-                                          callback(CMUserAccountPasswordChangeSucceeded, [NSArray array]);
-                                      }];
-                                  } else  {
-                                      callback(result, [NSArray array]);
-                                  }
-                              }
-     ];
-    
-//    [self changeUserCredentialsWithPassword:oldPassword newPassword:newPassword newUsername:nil newUserId:nil callback:callback];
-    
+    [self changeUserCredentialsWithPassword:oldPassword newPassword:newPassword newUsername:nil newUserId:nil callback:callback];
 }
 
 - (void)changeUserIdTo:(NSString *)newUserId password:(NSString *)currentPassword callback:(CMUserOperationCallback)callback {
@@ -367,20 +357,20 @@ static CMWebService *webService;
     [self changeUserCredentialsWithPassword:currentPassword newPassword:nil newUsername:newUsername newUserId:nil callback:callback];
 }
 
-//private?
 - (void)changeUserCredentialsWithPassword:(NSString *)currentPassword
                               newPassword:(NSString *)newPassword
                               newUsername:(NSString *)newUsername
                                 newUserId:(NSString *)newUserId
                                  callback:(CMUserOperationCallback)callback {
     
-    [webService updateCredentialsForUser:self
+    [webService changeCredentialsForUser:self
                                 password:currentPassword
                              newPassword:newPassword
                              newUsername:newUsername
                                newUserId:newUserId
                                 callback:^(CMUserAccountResult result, NSDictionary *responseBody) {
-                                    if (result == CMUserAccountCredentialsChangeSucceeded ||
+                                    
+                                    if (result == CMUserAccountCredentialChangeSucceeded ||
                                         result == CMUserAccountPasswordChangeSucceeded ||
                                         result == CMUserAccountUsernameChangeSucceeded ||
                                         result == CMUserAccountUserIdChangeSucceeded) {
@@ -397,6 +387,7 @@ static CMWebService *webService;
                                             self.userId = newUserId;
                                         }
                                         
+                                        // Only login if it was successful, otherwise it won't expire the session token.
                                         [self loginWithCallback:^(CMUserAccountResult resultCode, NSArray *messages) {
                                             callback(result, [NSArray array]);
                                         }];