+++ ios/ios/src/Web Services/CMUserAccountResult.h
@@ -29,16 +29,23 @@ typedef enum {
     /** The user Username change for a user succeeded */
     CMUserAccountUsernameChangeSucceeded,
     /** The user credentials change succeeded */
-    CMUserAccountCredentialsChangeSucceeded,
+    CMUserAccountCredentialChangeSucceeded,
     /** The forgotten password email was sent for the user */
     CMUserAccountPasswordResetEmailSent,
-
+    /** The credential change for a user failed due to an incorrect password/user */
+    CMUserAccountCredentialChangeFailedInvalidCredentials,
     /** Account creation failed because of an invalid email address or password */
     CMUserAccountCreateFailedInvalidRequest,
     /** The user profile update failed. See the accompanying dictionary for reasons. */
     CMUserAccountProfileUpdateFailed,
     /** Account creation failed because a user with that email address already exists for the current app */
     CMUserAccountCreateFailedDuplicateAccount,
+    /** The user credential change failed because the new User ID was already taken */
+    CMUserAccountCredentialChangeFailedDuplicateUserId,
+    /** The user credential change failed because the new Username was already taken */
+    CMUserAccountCredentialChangeFailedDuplicateUsername,
+    /** The user credential change failed because the new Username OR User ID was already taken */
+    CMUserAccountCredentialChangeFailedDuplicateInfo,
     /** The login failed due to an incorrect password for the given email address */
     CMUserAccountLoginFailedIncorrectCredentials,
     /** The password change for a user failed due to an incorrect password for the given email address */